Add manual reordering of accounts (#597)
Adds a "Reorder accounts" entry point beside the add-account button on
the Accounts screen, opening a dedicated drag-and-drop reorder screen
modeled on the existing category-group reorder flow.

Accounts already carried a sort_order column that Actual's CRDT layer
accepted writes for, but nothing in the app ever wrote to it. This adds
ActualEntityWriter.moveAccount (mirroring moveCategoryGroup) using the
same SortOrder.shove fractional-index algorithm, plus
AccountReorderPlanner/AccountDragReorder pure-logic helpers for the
drag gesture and up/down controls.

ActualAccount.sortOrder changes from Int to Double to match how
category/category-group sort_order is already modeled, since shove()
produces fractional values that would collide after a few reorders if
truncated to Int.

+++ app/src/main/java/com/azimulkabir/actua/data/budget/ActualBudgetDatabase.kt
@@ -103,7 +103,7 @@ class ActualBudgetDatabase private constructor(
                     type = ActualAccountType.fromDatabase(cursor.stringOrNull(2)),
                     offBudget = cursor.intOrZero(3) == 1,
                     closed = cursor.intOrZero(4) == 1,
-                    sortOrder = cursor.doubleOrZero(5).toInt(),
+                    sortOrder = cursor.doubleOrZero(5),
                     balanceCents = balances[id]?.total ?: 0,
                     clearedCents = balances[id]?.cleared ?: 0,
                     unclearedCents = (balances[id]?.total ?: 0) - (balances[id]?.cleared ?: 0),
